Query current time on message arrival instead of using can_msg::ts
Follows up b579638.

+++ libraries/YarpPlugins/CuiAbsolute/CuiAbsolute.hpp
@@ -246,7 +246,7 @@ protected:
 
     //-- Encoder stuff
     double encoder;
-    uint32_t encoderTimestamp;
+    double encoderTimestamp;
     yarp::os::Semaphore encoderReady;
     bool firstHasReached;
